refact: move min_spanning_tree benches to a different test file
