+++ benches/unionfind.rs
@@ -9,7 +9,7 @@ use test::Bencher;
 mod common;
 use common::*;
 
-use petgraph::algo::{connected_components, is_cyclic_undirected, min_spanning_tree};
+use petgraph::algo::{connected_components, is_cyclic_undirected};
 
 #[bench]
 fn connected_components_praust_undir_bench(bench: &mut Bencher) {
@@ -105,52 +105,4 @@ fn is_cyclic_undirected_petersen_dir_bench(bench: &mut Bencher) {
     let b = digraph().petersen_b();
 
     bench.iter(|| (is_cyclic_undirected(&a), is_cyclic_undirected(&b)));
-}
-
-#[bench]
-fn min_spanning_tree_praust_undir_bench(bench: &mut Bencher) {
-    let a = ungraph().praust_a();
-    let b = ungraph().praust_b();
-
-    bench.iter(|| (min_spanning_tree(&a), min_spanning_tree(&b)));
-}
-
-#[bench]
-fn min_spanning_tree_praust_dir_bench(bench: &mut Bencher) {
-    let a = digraph().praust_a();
-    let b = digraph().praust_b();
-
-    bench.iter(|| (min_spanning_tree(&a), min_spanning_tree(&b)));
-}
-
-#[bench]
-fn min_spanning_tree_full_undir_bench(bench: &mut Bencher) {
-    let a = ungraph().full_a();
-    let b = ungraph().full_b();
-
-    bench.iter(|| (min_spanning_tree(&a), min_spanning_tree(&b)));
-}
-
-#[bench]
-fn min_spanning_tree_full_dir_bench(bench: &mut Bencher) {
-    let a = digraph().full_a();
-    let b = digraph().full_b();
-
-    bench.iter(|| (min_spanning_tree(&a), min_spanning_tree(&b)));
-}
-
-#[bench]
-fn min_spanning_tree_petersen_undir_bench(bench: &mut Bencher) {
-    let a = ungraph().petersen_a();
-    let b = ungraph().petersen_b();
-
-    bench.iter(|| (min_spanning_tree(&a), min_spanning_tree(&b)));
-}
-
-#[bench]
-fn min_spanning_tree_petersen_dir_bench(bench: &mut Bencher) {
-    let a = digraph().petersen_a();
-    let b = digraph().petersen_b();
-
-    bench.iter(|| (min_spanning_tree(&a), min_spanning_tree(&b)));
 }
